refactor blacklist api
