+++ src/utils/api/index.js
@@ -1,4 +1,5 @@
 import Account from '@/utils/api/account.js';
+import Blacklist from '@/utils/api/blacklist.js';
 import Conversations from '@/utils/api/conversations.js';
 import Folders from '@/utils/api/folders.js';
 import Messages from '@/utils/api/messages.js';
@@ -6,6 +7,7 @@ import Stream from '@/utils/api/stream.js';
 
 export {
     Account,
+    Blacklist,
     Conversations,
     Folders,
     Messages,
